requirements.txt
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==17.0.0